sike: move common
* makes common reusable
* exports some more symbols from common
* remove kem for a moment
