+++ dh/sidh/common/doc.go
