+++ dh/sidh/common/types.go
