+++ dh/sidh/common/utils.go
