+++ dh/sidh/internal/p434/arith_test.go
@@ -6,7 +6,7 @@ package p434
 import (
 	"testing"
 
-	"github.com/henrydcase/nobs/dh/sidh/internal/common"
+	"github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Package-level storage for this field element is intended to deter
