+++ dh/sidh/internal/p434/params_test.go
@@ -4,7 +4,7 @@ package p434
 import (
 	"testing/quick"
 
-	. "github.com/henrydcase/nobs/dh/sidh/internal/common"
+	. "github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Values omputed using Sage
