+++ dh/sidh/internal/p503/arith_amd64_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"testing/quick"
 
-	"github.com/henrydcase/nobs/dh/sidh/internal/common"
+	"github.com/henrydcase/nobs/dh/sidh/common"
 	"golang.org/x/sys/cpu"
 )
