+++ dh/sidh/internal/p503/params_test.go
@@ -4,7 +4,7 @@ package p503
 import (
 	"testing/quick"
 
-	. "github.com/henrydcase/nobs/dh/sidh/internal/common"
+	. "github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Values omputed using Sage
