+++ dh/sidh/internal/p751/arith_test.go
@@ -6,7 +6,7 @@ package p751
 import (
 	"testing"
 
-	"github.com/henrydcase/nobs/dh/sidh/internal/common"
+	"github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Package-level storage for this field element is intended to deter
