+++ dh/sidh/internal/p751/params_test.go
@@ -4,7 +4,7 @@ package p751
 import (
 	"testing/quick"
 
-	. "github.com/henrydcase/nobs/dh/sidh/internal/common"
+	. "github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 var (
