+++ dh/sidh/internal/templates/arith_generic.gotemp
@@ -8,7 +8,7 @@ package {{ .PACKAGE}}
 import (
 	"math/bits"
 
-	"github.com/henrydcase/nobs/dh/sidh/internal/common"
+	"github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Compute z = x + y (mod p).
