+++ dh/sidh/internal/templates/curve.gotemp
@@ -4,7 +4,7 @@
 package {{ .PACKAGE}}
 
 import (
-	. "github.com/henrydcase/nobs/dh/sidh/internal/common"
+	. "github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Stores isogeny 3 curve constants
