+++ dh/sidh/internal/templates/fp2.gotemp
@@ -4,7 +4,7 @@
 package {{ .PACKAGE}}
 
 import (
-	"github.com/henrydcase/nobs/dh/sidh/internal/common"
+	"github.com/henrydcase/nobs/dh/sidh/common"
 )
 
 // Montgomery multiplication. Input values must be already
